requirements.txt
Jinja2==2.10
marshmallow==2.15.4
marshmallow-mongoengine==0.9.1
mongoengine==0.15.3
pymongo==3.6.0
python-decouple==3.1
redis==2.10.6
rq==0.11.0
tornado==5.0.2
tweepy==3.6.0
blinker==1.4
factory_boy==2.11.1